chore: update version to 0.0.6 and modify CHANGELOG; change valor type to double in PseTransactionData

+++ lib/data/models/pse_transaction_response.dart
@@ -35,7 +35,7 @@ class PseTransactionData {
   final int? refPayco;
   final String? factura;
   final String? descripcion;
-  final int? valor;
+  final double? valor;
   final int? iva;
   final int? baseiva;
   final String? moneda;
